added AbstractControllerSecurity
