+++ restservice/src/test/java/org/itevents/controller/AbstractControllerTest.java
@@ -1,12 +1,6 @@
 package org.itevents.controller;
 
-import org.itevents.model.User;
-import org.itevents.test_utils.BuilderUtil;
 import org.mockito.MockitoAnnotations;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 
@@ -26,25 +20,6 @@ public abstract class AbstractControllerTest {
 
     protected void initMvc(Object controller) {
         mockMvc = MockMvcBuilders.standaloneSetup(controller).build();
-    }
-
-    protected void authenticationGuest() {
-        User user = BuilderUtil.buildUserGuest();
-        authenticationUser(user);
-    }
-
-    protected void authenticationUser(User user) {
-        Authentication auth = new UsernamePasswordAuthenticationToken(
-                user.getLogin(),
-                user.getPassword(),
-                Arrays.asList(new GrantedAuthority() {
-                    @Override
-                    public String getAuthority() {
-                        return user.getRole().getName();
-                    }
-                })
-        );
-        SecurityContextHolder.getContext().setAuthentication(auth);
     }
 
 }
